backend/app/services/resources/service.py
"""
Resources service implementation
"""
from typing import List, Optional
from datetime import datetime
from loguru import logger
from sqlalchemy.orm import Session
from geoalchemy2 import functions as geo_func

from .models import Resource, ResourceType, ResourceStatus
from ...core.exceptions import NotFoundError, DatabaseError

# Placeholder - will be replaced with actual SQLAlchemy models
class ResourcesService:
    """Service for managing resources"""
    
    def __init__(self, db_session: Session):
        self.db = db_session
    
    async def list_resources(
        self,
        resource_type: Optional[ResourceType] = None,
        location: Optional[str] = None,
        limit: int = 100,
        skip: int = 0
    ) -> List[Resource]:
        """List all resources with optional filtering"""
        try:
            # TODO: Implement actual database query
            # For now, return mock data
            resources = [
                Resource(
                    id="1",
                    name="Dar es Salaam Water Treatment Plant",
                    type=ResourceType.WATER,
                    location="Dar es Salaam",
                    latitude=-6.7924,
                    longitude=39.2083,
                    capacity=500000,
                    current_level=350000,
                    status=ResourceStatus.OPERATIONAL,
                    description="Main water treatment facility"
                ),
                Resource(
                    id="2",
                    name="Muhimbili National Hospital",
                    type=ResourceType.HEALTH,
                    location="Dar es Salaam",
                    latitude=-6.7944,
                    longitude=39.2081,
                    capacity=1500,
                    status=ResourceStatus.OPERATIONAL,
                    description="Largest referral hospital"
                ),
            ]
            
            # Filter by type
            if resource_type:
                resources = [r for r in resources if r.type == resource_type]
            
            # Filter by location
            if location:
                resources = [r for r in resources if location.lower() in (r.location or "").lower()]
            
            return resources[skip:skip + limit]
            
        except Exception as e:
            logger.error(f"Error listing resources: {str(e)}")
            raise DatabaseError(f"Failed to list resources: {str(e)}")
    
    async def get_resource(self, resource_id: str) -> Optional[Resource]:
        """Get a resource by ID"""
        try:
            # TODO: Implement actual database query
            # Mock implementation
            if resource_id == "1":
                return Resource(
                    id="1",
                    name="Dar es Salaam Water Treatment Plant",
                    type=ResourceType.WATER,
                    location="Dar es Salaam",
                    latitude=-6.7924,
                    longitude=39.2083,
                    capacity=500000,
                    current_level=350000,
                    status=ResourceStatus.OPERATIONAL
                )
            return None
            
        except Exception as e:
            logger.error(f"Error getting resource {resource_id}: {str(e)}")
            raise DatabaseError(f"Failed to get resource: {str(e)}")
    
    async def create_resource(self, resource_data: dict) -> Resource:
        """Create a new resource"""
        try:
            # TODO: Implement actual database insert
            resource = Resource(
                **resource_data,
                id="new-id",  # Generate UUID
                created_at=datetime.utcnow(),
                updated_at=datetime.utcnow()
            )
            return resource
            
        except Exception as e:
            logger.error(f"Error creating resource: {str(e)}")
            raise DatabaseError(f"Failed to create resource: {str(e)}")
    
    async def get_nearby_resources(
        self,
        latitude: float,
        longitude: float,
        radius: float,
        resource_type: Optional[ResourceType] = None
    ) -> List[Resource]:
        """Get resources near a location using PostGIS"""
        try:
            # TODO: Implement PostGIS spatial query
            # Example query:
            # SELECT * FROM spatial.resources 
            # WHERE ST_DWithin(
            #     geometry,
            #     ST_MakePoint(longitude, latitude)::geography,
            #     radius * 1000
            # )
            
            # Mock implementation
            resources = await self.list_resources(resource_type=resource_type)
            # Filter by proximity (mock)
            return resources
            
        except Exception as e:
            logger.error(f"Error getting nearby resources: {str(e)}")
            raise DatabaseError(f"Failed to get nearby resources: {str(e)}")
