backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
pymongo==4.6.0
redis==5.0.1
httpx==0.25.2
aiohttp==3.9.1
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
cryptography==41.0.7
pyjwt==2.8.0
apscheduler==3.10.4
pandas==2.1.4
numpy==1.26.2
geopandas==0.14.1
shapely==2.0.2
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
python-dotenv==1.0.0
loguru==0.7.2
tenacity==8.2.3
